+++ comment-updater-postprocessing/build.gradle
@@ -9,6 +9,16 @@ dependencies {
     implementation 'com.google.code.gson:gson:2.6.2'
 }
 
+task runPostProcessing(dependsOn: runIde) {}
+
+gradle.taskGraph.whenReady { graph ->
+    if (graph.hasTask(runPostProcessing)) {
+        runIde.args "PostProcessing", "$dataset", "$output", "$config"
+        runIde.jvmArgs "-Djava.awt.headless=true"
+        runIde.maxHeapSize = "20g"
+    }
+}
+
 //task runPostProcessing(dependsOn: run) {}
 //
 //gradle.taskGraph.whenReady { graph ->
