+++ comment-updater-postprocessing/src/main/kotlin/org/jetbrains/research/commentupdater/SampleWriter.kt
@@ -3,6 +3,7 @@ package org.jetbrains.research.commentupdater
 import org.jetbrains.research.commentupdater.dataset.DatasetSample
 import java.io.File
 import com.google.gson.Gson
+import java.io.FileWriter
 import java.io.Writer
 
 class SampleWriter(output: File) {
@@ -16,7 +17,8 @@ class SampleWriter(output: File) {
 
     fun open() {
         outputFile = File(outputPath)
-        outputFile.writeText("[")
+        writer = FileWriter(outputFile)
+        writer.write("[")
     }
 
     fun close() {
